requirements.txt
# requirements.txt
streamlit
requests
openpyxl
python-dotenv
google-generativeai
mcp
faker
pandas
plotly
schedule
groq
